Add solution for problem `C - Prefix Enlightenment` Link: https://codeforces.com/contest/1290/problem/C Tags: dfs and similar, dsu, graphs, *2400 Ref: https://codeforces.com/contest/1290/submission/70072865

+++ codeforces/1290/C.cpp
@@ -22,8 +22,8 @@ const int OO = 1e6;
 
 struct DSU {
   int parent[N];
-  int cost[N];
-  int total_cost = 0;
+  Long cost[N];
+  Long total_cost = 0;
 
   void init() {
     for (int i = 0; i < N; ++i) {
